net/netns: bind to the default interface on iOS and macOS
This enables magicsock to bind to the default interface on iOS and macOS
like we do on other platforms.

Updates #3363
