+++ net/netns/netns_default.go
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-//go:build (!linux && !windows && !darwin) || (darwin && ts_macext)
-// +build !linux,!windows,!darwin darwin,ts_macext
+//go:build !linux && !windows && !darwin && !ios
+// +build !linux,!windows,!darwin,!ios
 
 package netns
